refactored auth page
