+++ App/Pages/Index.js
@@ -8,7 +8,7 @@ import Home from './Home'
 import Stories from './Stories'
 import ProfilePage from './ProfilePage'
 import AuthForm from './AuthForm'
-import AuthForm2 from './AuthForm2'
+
 
 const Index = (props) => {
   const { classes } = props;
@@ -22,7 +22,6 @@ const Index = (props) => {
           <Route path="/stories" component={Stories} />
           <Route path="/profile" component={ProfilePage} />
           <Route path="/auth" component={AuthForm} />
-          <Route path="/auth2" component={AuthForm2} />
       </Switch>
     </Grid>
   );
